chore: minor cleanups
