+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/irrationals/log/constructorTests.kt
@@ -15,26 +15,26 @@ fun runConstructorTests() {
     assertDivByZero { Log(1L, 10, isDivided = true) }
     assertDivByZero { Log(BigInteger.ONE, 10, isDivided = true) }
 
-    assertFailsWith<ArithmeticException>("Cannot calculate log of 0") { Log(ExactFraction.ZERO) }
-    assertFailsWith<ArithmeticException>("Cannot calculate log of 0") { Log(0) }
-    assertFailsWith<ArithmeticException>("Cannot calculate log of 0") { Log(0L) }
-    assertFailsWith<ArithmeticException>("Cannot calculate log of 0") { Log(BigInteger.ZERO) }
+    var error = "Cannot calculate log of 0"
+    assertFailsWith<ArithmeticException>(error) { Log(ExactFraction.ZERO) }
+    assertFailsWith<ArithmeticException>(error) { Log(0) }
+    assertFailsWith<ArithmeticException>(error) { Log(0L) }
+    assertFailsWith<ArithmeticException>(error) { Log(BigInteger.ZERO) }
 
-    assertFailsWith<ArithmeticException>("Cannot calculate log of negative number") { Log(-ExactFraction.TEN) }
-    assertFailsWith<ArithmeticException>("Cannot calculate log of negative number") { Log(-10) }
-    assertFailsWith<ArithmeticException>("Cannot calculate log of negative number") { Log(-10L) }
-    assertFailsWith<ArithmeticException>("Cannot calculate log of negative number") { Log(-BigInteger.TEN) }
+    error = "Cannot calculate log of negative number"
+    assertFailsWith<ArithmeticException>(error) { Log(-ExactFraction.TEN) }
+    assertFailsWith<ArithmeticException>(error) { Log(-10) }
+    assertFailsWith<ArithmeticException>(error) { Log(-10L) }
+    assertFailsWith<ArithmeticException>(error) { Log(-BigInteger.TEN) }
+    assertFailsWith<ArithmeticException>(error) { Log(ExactFraction(-4, 3)) }
 
-    assertFailsWith<ArithmeticException>("Cannot calculate log of negative number") {
-        Log(ExactFraction(-4, 3))
-    }
-
-    assertFailsWith<ArithmeticException>("Log base must be greater than 1") { Log(ExactFraction.TEN, -1) }
-    assertFailsWith<ArithmeticException>("Log base must be greater than 1") { Log(ExactFraction.TEN, 0) }
-    assertFailsWith<ArithmeticException>("Log base must be greater than 1") { Log(ExactFraction.TEN, 1) }
-    assertFailsWith<ArithmeticException>("Log base must be greater than 1") { Log(10, -1) }
-    assertFailsWith<ArithmeticException>("Log base must be greater than 1") { Log(10L, 0) }
-    assertFailsWith<ArithmeticException>("Log base must be greater than 1") { Log(BigInteger.TEN, 1) }
+    error = "Log base must be greater than 1"
+    assertFailsWith<ArithmeticException>(error) { Log(ExactFraction.TEN, -1) }
+    assertFailsWith<ArithmeticException>(error) { Log(ExactFraction.TEN, 0) }
+    assertFailsWith<ArithmeticException>(error) { Log(ExactFraction.TEN, 1) }
+    assertFailsWith<ArithmeticException>(error) { Log(10, -1) }
+    assertFailsWith<ArithmeticException>(error) { Log(10L, 0) }
+    assertFailsWith<ArithmeticException>(error) { Log(BigInteger.TEN, 1) }
 
     // ExactFraction
     // zero
